Split Instructions into README, part1, part2./
Unfortunately this is not very readable as a diff. README
gives an intro to Mondrian patterns and directs learners
to a suitable part1/2 or alternative material.

part I is the basic workshop for beginners.
part II is currently just a stub for the advanced learners' part yet to be written.